requirements.txt
pandas==1.4.3
matplotlib==3.5.2
numpy==1.23.0
scikit-learn==1.1.1
tqdm==4.64.0
albumentations==1.2.0
torchmetrics==0.9.2
clearml==1.6.1
pytorch-metric-learning==1.5.0
fastapi==0.78.0
uvicorn==0.18.2
python-multipart==0.0.5
torch --extra-index-url https://download.pytorch.org/whl/cu113
torchvision --extra-index-url https://download.pytorch.org/whl/cu113